requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

beautifulsoup4==4.9.3

bs4==0.0.1

certifi==2021.5.30

charset-normalizer==2.0.3; python_version >= '3.0'

idna==3.2; python_version >= '3.0'

python-dotenv==0.19.0

requests==2.26.0

soupsieve==2.2.1; python_version >= '3.0'

urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

